requirements.txt
flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
openai==0.28.1
werkzeug==2.3.7
pandas==2.1.1
numpy==1.24.3
plotly==5.17.0
python-dateutil==2.8.2
requests==2.31.0
gunicorn==21.2.0
flask-cors==4.0.0